// apps/designer/src/types/database.ts
export type Json =
  | string
  | number
  | boolean
  | null
  | { [key: string]: Json | undefined }
  | Json[]

export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "12.2.3 (519615d)"
  }
  public: {
    Tables: {
      account_credit_transactions: {
        Row: {
          account_id: string
          created_at: string
          credit_amount: number
          description: string | null
          id: string
          instance_id: string | null
          metadata: Json | null
          reload_attempt_description:
            | Database["public"]["Enums"]["credit_reload_attempt_description"]
            | null
          reload_attempt_status:
            | Database["public"]["Enums"]["credit_reload_attempt_status"]
            | null
          reload_type: Database["public"]["Enums"]["credit_reload_type"] | null
          type: Database["public"]["Enums"]["credit_transaction_type"]
        }
        Insert: {
          account_id: string
          created_at?: string
          credit_amount: number
          description?: string | null
          id?: string
          instance_id?: string | null
          metadata?: Json | null
          reload_attempt_description?:
            | Database["public"]["Enums"]["credit_reload_attempt_description"]
            | null
          reload_attempt_status?:
            | Database["public"]["Enums"]["credit_reload_attempt_status"]
            | null
          reload_type?: Database["public"]["Enums"]["credit_reload_type"] | null
          type: Database["public"]["Enums"]["credit_transaction_type"]
        }
        Update: {
          account_id?: string
          created_at?: string
          credit_amount?: number
          description?: string | null
          id?: string
          instance_id?: string | null
          metadata?: Json | null
          reload_attempt_description?:
            | Database["public"]["Enums"]["credit_reload_attempt_description"]
            | null
          reload_attempt_status?:
            | Database["public"]["Enums"]["credit_reload_attempt_status"]
            | null
          reload_type?: Database["public"]["Enums"]["credit_reload_type"] | null
          type?: Database["public"]["Enums"]["credit_transaction_type"]
        }
        Relationships: []
      }
// ...
      user_accounts: {
        Row: {
          account_id: string
          created_at: string
          id: string
          status: string
          updated_at: string
          user_id: string
          user_status: string
        }
        Insert: {
          account_id: string
          created_at?: string
          id?: string
          status?: string
          updated_at?: string
          user_id: string
          user_status?: string
        }
        Update: {
          account_id?: string
          created_at?: string
          id?: string
          status?: string
          updated_at?: string
          user_id?: string
          user_status?: string
        }
        Relationships: [
          {
            foreignKeyName: "user_accounts_account_id_fkey"
            columns: ["account_id"]
            isOneToOne: false
            referencedRelation: "accounts"
            referencedColumns: ["id"]
          },
        ]
      }
// ...
    }
// ...
  }
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof DatabaseWithoutInternals },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never